Fix OrganizationRepo
