+++ kifi-backend/modules/shoebox/app/com/keepit/model/OrganizationRepo.scala
@@ -1,13 +1,11 @@
 package com.keepit.model
 
 import com.google.inject.{ Provider, ImplementedBy, Inject, Singleton }
-import com.keepit.common.db.{ SequenceNumber, State, Id }
+import com.keepit.common.db.{ Id }
 import com.keepit.common.db.slick.DBSession.RSession
 import com.keepit.common.db.slick._
 import com.keepit.common.logging.Logging
-import com.keepit.common.mail.EmailAddress
 import com.keepit.common.time.Clock
-import org.joda.time.DateTime
 
 @ImplementedBy(classOf[OrganizationRepoImpl])
 trait OrganizationRepo extends Repo[Organization] with SeqNumberFunction[Organization] {
@@ -31,38 +29,7 @@ class OrganizationRepoImpl @Inject() (val db: DataBaseComponent, val clock: Cloc
     def organizationHandle = column[Option[OrganizationHandle]]("organization_handle", O.Nullable)
     def normalizedOrganizationHandle = column[Option[OrganizationHandle]]("normalized_organization_handle", O.Nullable)
 
-    def applyFromDbRow(
-      id: Option[Id[Organization]],
-      createdAt: DateTime,
-      updatedAt: DateTime,
-      state: State[Organization],
-      seq: SequenceNumber[Organization],
-      name: String,
-      description: Option[String],
-      ownerId: Id[User],
-      organizationHandleOpt: Option[OrganizationHandle],
-      normalizedOrganizationHandleOpt: Option[OrganizationHandle]) = {
-      val primaryOrganizationHandle = for {
-        organizationHandle <- organizationHandleOpt
-        normalizedOrganizationHandle <- normalizedOrganizationHandleOpt
-      } yield PrimaryOrganizationHandle(organizationHandle, normalizedOrganizationHandle)
-      Organization(id, createdAt, updatedAt, state, seq, name, description, ownerId, primaryOrganizationHandle)
-    }
-
-    def unapplyToDbRow(org: Organization) = {
-      Some((org.id,
-        org.createdAt,
-        org.updatedAt,
-        org.state,
-        org.seq,
-        org.name,
-        org.description,
-        org.ownerId,
-        org.handle.map(_.original)),
-        org.handle.map(_.normalized))
-    }
-
-    def * = (id.?, createdAt, updatedAt, state, seq, name, description, ownerId, organizationHandle, normalizedOrganizationHandle) <> ((applyFromDbRow _).tupled, unapplyToDbRow _)
+    def * = (id.?, createdAt, updatedAt, state, seq, name, description, ownerId, organizationHandle, normalizedOrganizationHandle) <> ((Organization.applyFromDbRow _).tupled, Organization.unapplyToDbRow _)
   }
 
   def table(tag: Tag) = new OrganizationTable(tag)
